// uiug-astro2026/src/api/types.ts
export interface paths {
    "/umbraco/delivery/api/v2/content": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["GetContent2.0"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/umbraco/delivery/api/v2/content/item/{path}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["GetContentItemByPath2.0"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/umbraco/delivery/api/v2/content/item/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["GetContentItemById2.0"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/umbraco/delivery/api/v2/content/items": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["GetContentItems2.0"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/umbraco/delivery/api/v2/media": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["GetMedia2.0"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/umbraco/delivery/api/v2/media/item/{path}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["GetMediaItemByPath2.0"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/umbraco/delivery/api/v2/media/item/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["GetMediaItemById2.0"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/umbraco/delivery/api/v2/media/items": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["GetMediaItems2.0"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        ApiLinkModel: {
            readonly url?: string | null;
            readonly queryString?: string | null;
            readonly title?: string | null;
            readonly target?: string | null;
            /** Format: uuid */
            readonly destinationId?: string | null;
            readonly destinationType?: string | null;
            readonly route?: components["schemas"]["ApiContentRouteModel"] | null;
            linkType: components["schemas"]["LinkTypeModel"];
        };
// ...
        SiteSettingsPropertiesModel: components["schemas"]["COm_siteSettingsPropertiesModel"] & components["schemas"]["COm_footerPropertiesModel"] & components["schemas"]["COm_headerPropertiesModel"] & components["schemas"]["COm_socialLinksPropertiesModel"] & components["schemas"]["COm_scriptsPropertiesModel"];
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}